Allow port parameter to be empty
Although parseInt('') returns NaN, isNaN('') is false. If the port
parameter is set to a blank string, it will thus be forced to 0 by the
Options.set method. This patch changes the test such that the set
method will allow options.port to be set to an empty value.

Since the browser helpfully will ignore the port field in an URI if no
actual number is specified, we can instruct the livereload script to
open a websocket on the same server (domain and port) that the script
came from.

This in turn is useful if you have the livereload server running on an
insecure port on a local interface, while the pages are served through
a proxy that handles SSL, as secure pages cannot contact insecure
websockets. By routing the livereload socket through the same server
as the hosting web page on the other hand, the proxy can do the
encryption for this traffic as well.