Let empty port value mean same port as script
If we specify '?port=' as parameter list (i.e. an empty port value),
this should mean that we want to run not on any particular other port,
but rather the same port as the injected script, as this is the only
way we can specify such a situation when also running on a non-default
port (for instance 8000 or 8080).

Since we couldn't specify empty ports before, this change shouldn't
affect any existing configurations, only enable new ones.