Keep port parameter numeric if possible, or string
Falsy values that are put in the port parameter (null, undefined,
blank string) are all mapped to a blank string, as if the port was
specified in the URL but with value omitted. If a string is passed,
but it can be interpreted as a number with the plus operator, then it
is kept as a number; this is the default behaviour.