derive port from host when blank (closes #98)
* Allow port parameter to be empty

Although parseInt('') returns NaN, isNaN('') is false. If the port
parameter is set to a blank string, it will thus be forced to 0 by the
Options.set method. This patch changes the test such that the set
method will allow options.port to be set to an empty value.

Since the browser helpfully will ignore the port field in an URI if no
actual number is specified, we can instruct the livereload script to
open a websocket on the same server (domain and port) that the script
came from.

This in turn is useful if you have the livereload server running on an
insecure port on a local interface, while the pages are served through
a proxy that handles SSL, as secure pages cannot contact insecure
websockets. By routing the livereload socket through the same server
as the hosting web page on the other hand, the proxy can do the
encryption for this traffic as well.

* Let empty port value mean same port as script

If we specify '?port=' as parameter list (i.e. an empty port value),
this should mean that we want to run not on any particular other port,
but rather the same port as the injected script, as this is the only
way we can specify such a situation when also running on a non-default
port (for instance 8000 or 8080).

Since we couldn't specify empty ports before, this change shouldn't
affect any existing configurations, only enable new ones.

* Fission test to get part that will be unmodified

* Update test-cases for empty port parameter behaviour

* Document the meaning of an empty port parameter

* Give port its own setter to handle blank values

Instead of amending the generic handler with code to handle blank
value, a setter for the port property alone is added, that handles
this particular case (since a blank value here have a special meaning;
that the port specifier should be dropped altogether in the backend
url, to accomodate proxies).

* Keep port parameter numeric if possible, or string

Falsy values that are put in the port parameter (null, undefined,
blank string) are all mapped to a blank string, as if the port was
specified in the URL but with value omitted. If a string is passed,
but it can be interpreted as a number with the plus operator, then it
is kept as a number; this is the default behaviour.

---------

Co-authored-by: Roland Kaufmann <rlndkfmn+livereload@gmail.com>